requirements.txt
# Web scraping and HTTP requests
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and Supabase
supabase==2.0.2
psycopg2-binary==2.9.7
asyncpg==0.29.0

# Data validation and serialization
pydantic==2.5.0
pydantic[email]==2.5.0

# Async support
aiofiles==23.2.1
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0

# Date and time handling
python-dateutil==2.8.2

# Logging and monitoring
loguru==0.7.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools (optional)
black==23.11.0
flake8==6.1.0
isort==5.12.0